add illustration with plt show

+++ model/metrics.py
@@ -4,7 +4,7 @@ import numpy as np
 
 def vi(mask: np.array, pred: np.array):
     mask_label = label(mask, background=0, connectivity=1)
-    pred_label = label(pred > 0.5, background=0, connectivity=1)
+    pred_label = label((pred > 0.5).astype(int), background=0, connectivity=1)
     merger_error, split_error = metrics.variation_of_information(pred_label, mask_label)
 
     return merger_error + split_error
